+++ packages/charts-figma/src/DisplayChart.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2022 Visa, Inc.
+ * Copyright (c) 2022, 2024 Visa, Inc.
  *
  * This source code is licensed under the MIT license
  * https://github.com/visa/visa-chart-components/blob/master/LICENSE
@@ -21,7 +21,7 @@ import {
   DumbbellPlot,
   WorldMap,
   AlluvialDiagram
-} from './visa-charts';
+} from '@visa/charts-react';
 
 let chartProps = {};