+++ packages/keyboard-instructions/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.3](https://github.com/visa/visa-chart-components/compare/@visa/keyboard-instructions@1.3.1...@visa/keyboard-instructions@1.3.3) (2024-02-16)
+
+**Note:** Version bump only for package @visa/keyboard-instructions
+
 ## [1.3.2](https://github.com/visa/visa-chart-components/compare/@visa/keyboard-instructions@1.3.1...@visa/keyboard-instructions@1.3.2) (2024-02-09)
 
 **Note:** Version bump only for package @visa/keyboard-instructions